dankomax - added TUA-437 to API ChallengeTest

+++ src/main/java/com/ita/edu/speakua/api/clients/ChallengeClient.java
@@ -13,7 +13,7 @@ public class ChallengeClient extends BaseClient {
         this.authentication = authentication;
     }
 
-    public Response get(int id) {
+    public Response get(long id) {
         return prepareRequest()
                 .header("Authorization", "Bearer " + this.authentication)
                 .when()
